implement searchform and metrics
